+++ lib/yandex_cloud.rb
@@ -1,4 +1,6 @@
+require 'dotenv/load'
 require_relative 'yandex_cloud/version'
+require_relative 'yandex_cloud/auth'
 
 # define main endpoint
 module YandexCloud