+++ CHANGELOG.md
@@ -7,3 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## Unreleased
 ### Added
 - CHANGELOG
+- dotenv support
+- YandexCloud::Auth with token()
+- specs for YandexCloud::Auth.token()
+
+### Modified
+- README